change: move widget to its own class, check field name instead of type
